refactor(structure): relocate planner calc utils into utils/planner
