+++ src/utils/planner/planScorer.ts
